requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Streamlit Frontend
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2

# Supabase
supabase==2.0.3
postgrest==0.13.0

# AI/ML
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2

# Data Processing
yfinance==0.2.32
requests==2.31.0
python-dotenv==1.0.0

# Prometheus
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
aiohttp==3.9.1

